move apiValidation to utils
